Alter the reset button UI.
There are now 5 states: rest, select, grace, commit and abort.

The device is in the rest state until the button is depressed. It then
moves to the select state.

In the select state, the LED cycles between the following modes:

- halt (grey)
- reboot (green)
- reset-userdata (yellow)
- reset-root (red)
- abort (white)

When the user releases the button, the device enters the grace state for the specified mode unless the selection
was abort, in which case it enters the abort state.

In the grace state, the color fades. During this time, if the user
presses the button again, the device moves into the abort state. Otherwise,
the device proceeds to the commit state and the action selected action
is taken.

In the abort state, the color fades from white to black and then the device returns to the rest state.


Signed-off-by: Jon Seymour <jon@ninjablocks.com>